requirements.txt
apyori==1.1.2
kneed==0.8.1
matplotlib==3.6.2
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
scikit_learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
streamlit==1.16.0
streamlit_player==0.1.5
